+++ feed/edit.html
@@ -1,0 +1,50 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description" content="Edit your post on Social Media where you can find out what your friends are up to." />
+    <title>Post | Social Media</title>
+    <script src="https://kit.fontawesome.com/d2a33a8755.js" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="/dist/css/index.css">
+    <script src="/js/feed/edit.mjs" type="module" defer></script>
+</head>
+<body class="m-0">
+    <header class="container text-center align-items-center min-vw-100 h-100 p-2">
+        <div class="row">
+            <div class="col input-group w-25 h-25 my-auto">
+                <span class="input-group-text"><i class="fa-solid fa-magnifying-glass"></i></span>
+                <input type="search" name="search" class="form-control" autocomplete="off" placeholder="">
+            </div>
+            <h1 class="col-6 fs-2 my-auto"><a href="index.html">Social Media</a></h1>
+            <div class="col dropdown">
+                <img src="/images/avatar-1295395_640.png" alt="penguin profile image" class="dropdown-toggle w-25" role="button" data-bs-toggle="dropdown">
+                    <ul class="dropdown-menu">
+                        <li><a href="/profile/index.html" class="dropdown-item">Profile</a></li>
+                        <li class="dropdown-item" id="logOut">Log out</li>
+                    </ul>
+            </div>
+        </div>
+    </header>
+    <main>
+        <div id="post_container">
+            <div class="panel d-flex flex-column justify-content-center w-50 m-auto my-4">
+                <div class="panel-body">
+                    <label for="post" class="mb-2">Create new post</label>
+                    <input class="form-control mb-2" name="title" id="post_title" rows="1" placeholder="">
+                    <textarea class="form-control" name="body" id="post_body" rows="2" placeholder=""></textarea>
+                    <input class="form-control mt-2" name="tags" id="post_tags" rows="1" placeholder="Add some tags if you wish." >
+                    <div class="d-flex align-items-center my-2 p-2 gap-2">
+                        <i class="fa-regular fa-image fs-2"></i>
+                        <input class="form-control-sm" type="url" name="media" id="post_image" placeholder="Add a link to the picture.">
+                    </div>
+                    <button id="updateBtn" class="btn btn-sm btn-secondary px-4 py-2 d-flex justify-content-center">Update</button>
+                    <a href="../feed/index.html"><button class="btn btn-sm btn-secondary px-4 py-2 d-flex justify-content-center mt-2">Cancel</button></a>
+                </div>
+            </div>
+        </div>
+    </main>
+</body>
+</html>
