+++ js/feed/edit.mjs
@@ -4,7 +4,10 @@ import { putMethod } from "../imports/request-methods/put.mjs";
 
 const title = document.querySelector("#post_title");
 const body = document.querySelector("#post_body");
-const image = document.querySelector("#post_image");
+const image = document.querySelector("#post_image")
+const tags = document.querySelector("#post_tags");
+
+const updateBtn = document.querySelector("#updateBtn");
 
 
 const queryString = document.location.search;
@@ -13,16 +16,32 @@ const id = params.get("post");
 
 const postURL = urlPosts + "/" + id;
 
-
 export async function editPost(data) {
     try {
-        console.log(data)
         const putRequest = await putMethod(data);
         const response = await fetch(postURL, putRequest);
         const json = await response.json();
-        console.log(json)
+
+        if(response.ok) {
+            window.location.replace(`/feed/post.html?post=${id}`);
+        } else {
+            console.log("You can't update this post!");
+        }
 
     } catch (error) {
         console.log(error);
     }
 }
+
+const imgSrc = document.getElementById("post_image").getAttribute("src");
+
+updateBtn.addEventListener("click", () => {
+    const request =  {
+        title: title.value,
+        body: body.value,
+        media: image.value,
+        tags: tags.value.split(",")
+    }
+    console.log(title)
+    editPost(request);
+})
